Super optimization by using a different data type for the operations Had to do some math wizardry for it to work. Today started with 250 bytes/sec, now it's 2500 bytes/sec, and on my PC, which is somehow even slower than my laptop.

+++ ReedSalomon/src/ReedSalomon.h
@@ -19,8 +19,9 @@
 #include "CommonDefines.h"
 
 // You can specify the type of the numbers used by the algorithm. This will be related to MODULUS as
-// the type you select has to be able to represent at least (MODULUS-1)^2.
-typedef int ModInt;
+// the type you select has to be able to represent at least (MODULUS-1)^2. There are no negative 
+// values, so unsigned values can be used.
+typedef unsigned short ModInt;
 
 /***************************************************************************************************
  * ALGORITHM SELECTION
